New "Airport Finder" dialog
* This dialog allows one to find all airports located in a given area
specified with maximum and minimum distances to a reference airport.
The direction of the path (from or to the reference airport) can be
chosen. For each result, the dialog gives the orthodromic distance
(i.e., the length of the shortest path following the Earth curvature)
between the two airports as well as the initial and final bearings of
the course, according to the chosen direction. It is also possible to
choose between magnetic and true bearings, nautical miles and kilometers
(for display of the results), Vincenty and Karney calculation methods.
Finally, a button allows one to choose the airport selected in the
results table and automatically close the dialog.