Fix calculation of airport centroid coordinates
* ffgo/fgdata/apt_dat.py (AptDat.makeAptDigest, AptDat._readAirportData,
AptDat._processPotentialRunwayRow): use n-vectors to compute the
latitude and longitude of the centroid of each airport. The previous
method, a simple arithmetic mean, could give incorrect results with
airports having runway ends located on either side of the 180° meridian
or very close to the North or South pole (not that I encountered any
such problem, but I prefer a correct method, all the more since the
performance impact is negligible).