Add the minimum and maximum runway lengths to the apt digest file
* Helipad lengths, even though they are stored in the 'length' attribute
of Airport and AirportStub instances, are not considered at all in the
calculation of minimum and maximum runway lengths for a given airport.

* The result can be found after startup in the AirportStub instances
that are Config.airports' values. If an airport has no runway (it should
be a heliport), then AirportStub.minRwyLength and
AirportStub.maxRwyLength are both set to None. Otherwise, both of these
attributes are floats and should be interpreted as lengths in meters, as
the runway length and width.

* Don't increase the format version number of the apt digest file, as
the previous commit hasn't been published yet.