Airport Finder: allow resizing the three main LabelFrame widgets
* Replace the top-level 'tlFrame' Frame with a PanedWindow. Enclose each
of the three LabelFrame widgets inside a Frame that allows fine control
of the padding on each side of the LabelFrame widget
('refAirportOuterFrame', 'searchParamsOuterFrame', 'resultsOuterFrame').
Alternatively, one could do with the LabelFrame widgets directly inside
the PanedWindow, itself directly inside a unique Frame inside the
Toplevel, but that way wouldn't allow to control the vertical separation
between the LabelFrame widgets.

* Each of the three main parts of the Airport Finder dialog ('Reference
airport', 'Search parameters' and 'Search results') can now be resized
with the mouse.