Fix wrong line number in error reports for index-based apt.dat requests
* ffgo/fgdata/apt_dat.py (AptDat.readAirportDataUsingIndex): this method
used to seek into the apt.dat.gz file using the specified index, but it
did not modify AptDat.lineNb accordingly. As a consequence, the line
numbers for error reports concerning apt.dat parsing raised by this
method were usually wrong.

* In order to fix this, the line number is stored along with the byte
offset in the apt digest file. Since both of these are closely related,
this commit changes the meaning of "index" in the code: before this
commit, it referred to a byte offset in apt.dat; now, it refers to a
tuple (byteOffset, lineNb) where 'byteOffset' points to the start of the
first line defining an airport in apt.dat, and 'lineNb' is the
corresponding line number (starting from 1). In other words, it contains
the same byte offset as before, but inside a tuple that also contains
the line number corresponding to this byte offset. This way, both tend
to stay together (often, they are just passed along) and changes to the
code are minimal.

* Unfortunately, in order to store the line number along with the
corresponding byte offset in the apt digest file, it is necessary to
modify its format a little bit. This commit thus changes the apt digest
file format version from 2 to 3.