Remove GUI code from ffgo/fgdata/parking.py
* ffgo/fgdata/parking.py: this file is supposed to be usable
independently of any GUI. The previous commit introduced Tkinter in this
module to fix a little bug. This commit changes
ffgo.fgdata.parking.readGroundnetFile() to return a tuple containing the
usual result, as well as a list of the exceptions caught while parsing
the groundnet file.

* ffgo/gui/mainwindow.py: adapt the two calls to
ffgo.fgdata.parking.readGroundnetFile(), and change the second one to be
more consistent with the first one.