Use widgets.AirportChooser for the airport list in the main window
* The airport list in the main window used to be a simple list; replace
it with a real multicolumn widget (Ttk Treeview), reusing the logic
already implemented in widgets.AirportChooser. This fixes small
alignment problems in the previous implementation (due to the ICAO codes
not having the same display width), allows one to sort the airport
list by ICAO code or airport name, in ascending or descending order,
with a simple click on the relevant column header, and brings the nice
logic from widgets.AirportChooser to the main window airport list (such
as preferring an exact match on ICAO code when there are also substring
matches on airport names for a given search string).

* In order to make things more intuitive and robust, the semantics of
Config.airport and Config.carrier have been slightly changed: whereas
carrier mode used to be marked by Config.airport.get() being empty, the
condition to check whether we are in “carrier mode” is now
“Config.carrier.get() is not empty”. It is possible to have both
Config.airport.get() and Config.carrier.get() empty at the same time:
this happens when the airport search text matches no airport while the
user is not in carrier mode.

Besides being well-defined and intuitive, the new semantics allow the
program to remember the selected airport while in carrier mode, and
to restore it afterwards in a straightforward way.