Work around the main window layout problems after saving the Preferences
* Workaround for a weird problem: when saving the settings from the
Preferences dialog, even if the *very same font size* is set as the one
that was used at program initialization, the main window layout gets
broken, with the airport chooser Treeview taking more and more
horizontal space every time the settings are saved. Work around the
problem by avoiding to reconfigure the fonts in such "reinit" conditions
(i.e., when exiting the Preferences dialog with the "Save" button).

* The new font size---assuming this parameter was changed---is still set
internally in the Config instance and written to the configuration file
as soon as "Save" is clicked, therefore the essential drawback of this
workaround is that font size changes are not immediately visible anymore
upon exit from the Preferences dialog. In such a case, display a message
telling the user that he should maybe restart FFGo to see the effects of
the font size change.