New GPS Tool dialog
* ffgo/gui/gps_tool.py: new dialog allowing one to find the distance,
initial and final bearings for the shortest path between two given
airports. This dialog should be more convenient than the Airport Finder
dialog for the cases where you already know the start and destination
airports you are interested in, and you just want to get the results for
these airports. Or you want to check several pairs of airports, but you
know *which* ones you want the calculations to be made for. The dialog
also allows one to see the flight duration for a given ground speed, and
vice versa. It reuses the AirportChooser widget already used in FFGo's
main window and in the Airport Finder dialog.

* ffgo/gui/mainwindow.py: integrate the new dialog.

* ffgo/config.py: call the new module's
setupEarthMagneticFieldProvider() function from
Config.updateMagFieldProvider().