Airport Finder: hide rather than destroy the Toplevel window
* Similarly as for the GPS Tool dialog, if the Airport Finder dialog is
recreated every time it is opened, then memory occupation increases
every time the dialog is opened and does not decrease when it is
destroyed (tested on Linux). On the contrary, if the dialog is hidden
with withdraw() and shown again with deiconify(), then memory occupation
does not increase once the dialog has been created. This also has the
advantage of making the dialog appear more promptly on the second and
subsequent open operations. Implement this behavior for the Airport
Finder. (maybe one would need to explicitely destroy() every single
widget to actually see memory decrease with the previous "destroy()"
strategy...)

* ffgo/gui/airport_finder.py (AirportFinder.quit): rename this method
(now unused) to 'destroy()' for clarity, as opposed to 'hide()'.