Fix clearing of the airport list in App.selectNewAirport()
* ffgo/gui/mainwindow.py (App.selectNewAirport): replace
self.airportSearch.delete(0, 'end') with
self.airportChooser.clearSearch(), which is there just for this purpose
and reflects the change immediately.

* ffgo/gui/mainwindow.py (App.selectNewAirport): call
self.master.update_idletasks() after the
self.airportChooser.clearSearch() call in order to let Tk update the
airport list, and before calling
self.airportList.FFGoGotoItemWithValue(). Otherwise, if the airport
search text is not empty, it may be that the desired airport is not
currently displayed, and thus can't be selected.