Rebase AirportChooser on a generic, abstract class: IncrementalChooser
* ffgo/gui/widgets.py: strip everything that is airport-specific out of
the AirportChooser class and rename it to IncrementalChooser. The new
class is an abstract class that can be easily derived to build many
kinds of multicolumn lists with a search field and a Clear button. It
will be used for the new aircraft chooser in future commits.

* ffgo/gui/widgets.py: redefine the AirportChooser class as a subclass
of IncrementalChooser (basically: provide an implemention for all
abstract methods of IncrementalChooser).

* Other files: improve readability of AirportChooser-creating calls.