Reimplement the aircraft chooser using widgets.IncrementalChooser
* ffgo/gui/widgets.py (AircraftChooser): new class, derived from
IncrementalChooser. Use this new class to implement the aircraft chooser
in FFGo's main window. This improves its behavior in several respects:
  - smarter automatic selection of an aircraft when the search query is
    changed, based on the new search query and the previous selection;
  - better handling of navigation keys (up and down arrows, Page Up and
    Page Down, Home and End; most notably, holding down one of the first
    four of these, and pressing one of the last two should have the
    expected effect);
  - ability to display several proper columns of data and sort by
    clicking on a column header;
  - short delay before starting to search after the search query is
    modified, in order to allow smooth typing in the search field.

For now, the aircraft chooser has only one column, but another one
should be added soon.

* Revamp the way the current aircraft is stored in the Config class and
managed in particular from the App class. The central variable telling
which is the current aircraft is Config.aircraftId (an instance of
misc.Observable). Its value is a tuple of the form
'(aircraftName, aircraftDir)'. It is automatically updated by the
AircraftChooser instance used for the aircraft chooser in FFGo's main
window (Config.aircraftId is the output variable of this AircraftChooser
instance). Whenever it is set, Config.aircraft and Config.aircraftDir
are automatically updated via an observer, as well as the aircraft
thumbnail and the FlightGear command (two other observers). This new
method of managing the aircraft selection reduces the amount of code in
the App class. It should also be easier to understand and keep
consistent.

* ffgo/gui/mainwindow.py (App): remove useless (redundant) calls to
App.updateImage().