widgets.IncrementalChooser: make 'searchVar' an instance of misc.Observable
* ffgo/gui/widgets.py (IncrementalChooser): since 'searchVar' is not
directly linked to any Tk widget (the one that is linked to the search
Entry is 'searchBufferVar'), it can be transparently changed into an
instance of misc.Observable. This should improve performance a tiny bit
(maybe not noticeable) and in some cases ease debugging. It also
increases FFGo's portability to another GUI toolkit...