IncrementalChooser: autoselect exact matches even when tree contents unchanged
* ffgo/gui/widgets.py (IncrementalChooser): when 'searchVar' is modified
(search query change becoming effective), updateList() is called. This
used to cause the tree widget to be updated *if* the underlying data
('matches') had changed due to the new search query. After updating the
tree widget in such a case, a suitable item was automatically selected,
but this didn't happen if the new search query had not changed the tree
contents (the set of matches). This is incorrect, because it is possible
to obtain an exact match by making a change to the search query that
doesn't modify the tree contents.

In order to fix this, move the automatic item selection previously done
in _updateTreeWidget() to a new method (_autoUpdateTreeSelection()), and
call this method from updateList(), regardless of whether the list of
matches has changed due to the new search query.

* ffgo/gui/widgets.py (IncrementalChooser.__init__): new optional
keyword argument 'clearSearchOnInit' defaulting to True. Setting it to
False prevents IncrementalChooser.__init__() from calling clearSearch()
at the end of its execution.

* ffgo/gui/mainwindow.py (App.__init__): initialize the airport and
aircraft choosers with clearSearchOnInit=False, otherwise
self.config.airport and self.config.aircraftId would be nullified at
application initialization, since no item could possibly be selected in
the empty tree widgets (remember that the airport and aircraft lists are
only filled later, in reset()).