Aircraft chooser: approximate matching of aircraft names
* With this commit, you don't have to remember that the 'f-14b' aircraft
has a hyphen, the 'EC-137D' also, but not the 'ec130b4' and 'ec135p2'.
Characters in the following string are simply ignored during the first
phase (filtering) of an aircraft search: ' -_.,;:!?' (the list starts
with a space). This means that an aircraft search query can find an
aircraft even if it has hyphens, underscores, etc. missing or in excess
compared to the actual aircraft name.

* Note that this only concerns the aircraft list *filtering*. Given a
set of aircrafts that match the search query, one of them may be
automatically selected after the aircraft list is filtered if:
  - it is either an exact (case-insensitive) match of the search query
    (and for this purpose, characters such as hyphens and underscores
    are *not* ignored: autoselection requires an exact match);
  - or it is the previously-selected aircraft (option only taken if the
    first one couldn't be satisfied, and of course if the
    previously-selected aircraft is in the just-updated, filtered list)

* As said, both types of matching are case-insensitive:
  1) the approximate matching process used to determine which aircrafts
     will be displayed in the aircraft list after a change to the search
     field;
  2) the exact matching process used to autoselect the "best" among the
     matches found in 1), falling back to the first one if none of these
     fulfills any of the two conditions given in the previous paragraph.