Pass --fg-root when running 'fgfs --version' to avoid the new stupid popup
* ffgo/fgdata/fgversion.py (getFlightGearVersion): FlightGear 2016.1.0
(from February 2016) can spawn an annoying popup dialog when
'fgfs --version' is run, which is very confusing for users and breaks
FFGo's detection of the FlightGear version (this popup is apparently
meant to let the user graphically choose the FG_ROOT path to use in the
built-in Qt lancher). Pass the --fg-root option as well (assuming
Config.FG_root is non-empty, otherwise FlightGear version detection is
not attempted) to work around this new behavior of 'fgfs --version'.

getFlightGearVersion() takes a new argument: the value to pass to
fgfs' --fg-root option.

* ffgo/config.py (Config.getFlightGearVersion): pass the additional
required argument if Config.FG_root is non-empty, otherwise don't
attempt to detect the FlightGear version.