Just use "Aircraft" instead of "Aircraft name" for the ac chooser col title
* ffgo/gui/mainwindow.py (App.__init__): the "Aircraft name" column
title in the aircraft chooser is uselessly long and not so accurate,
depending on whether one is thinking of the real aircraft name or its
name in FlightGear -> just use "Aircraft"; the FFGo main window has
enough stuff to read in it already that we don't need to add more.

* Still leaving "Airport name" as the column title in the airport
chooser, as in this case, it is not some abbreviation or whatever, it
really gives the airport name.