Preferences dialog: use the grid() layout manager instead of pack()
* ffgo/gui/configwindow.py: reimplement the whole dialog using the
grid() layout manager instead of pack(). This allows to have proper
horizontal and vertical alignments in the recently-added "Statistics"
notebook pane. This should also make it easier to add things or make
changes to the dialog in the future.

* ffgo/gui/configwindow.py: get rid of the wilcard import
'from tkinter import *' in favor of the much less intrusive
'import tkinter as tk'.