Update the translated template config files ffgo/data/config/config_*
* Remove the --multiplay option: nowadays, most people should connect
using the in-sim menu when they are ready (not on a runway, not on
someone else, and ready to answer an eventual ATC operator, for one).
Other people can find the option on the FlightGear wiki anyway.

* Update the French template: should now be up-to-date except for
--terrasync-dir which is not mentioned (the option may cause problems
due to new permission restrictions in current FlightGear from Git next,
but this is still in flux so, for now, don't translate this).