Harmonize spacing before commented out options in ffgo/data/config/config_*
* The old way '# --option=value' instead of '#--option=value' was
probably more readable, but the sentence at the top of each file only
says to remove the '#' character to activate an option, not '# '. That
could certainly be changed, but given that the only maintained
translations are French and German, I can't make such a change now
without breaking all other translations. (yes, having an uncommented
option line starting with a space works, but it is ugly...)