Pressure converter: rename things
* ffgo/gui/pressure_converter.py: remove the assumption that an
"altimeter setting" is always in inHg and a "QNH" is always in hPa. I
had that idea when writing the previous commits, but I now believe this
was a misconception due to habit (ATC sessions...), and that a "QNH" may
very well be expressed in both units without bending the meaning of this
term.

* Many lines are affected because of variable and function/method names,
but this is just a dumb renaming throughout the whole file. Basically:

  altimeter setting -> renamed to inHg
  QNH               -> renamed to hPa