Use accurate wording to refer to GeographicLib's Python 3 implementation
* Consistently use "GeographicLib's implementation for Python 3", or
maybe "GeographicLib's Python 3 implementation" when space is scarce,
but not wordings such as "GeographicLib's binding for Python 3". This is
because 'pip install geographiclib' does *not* install a binding to the
GeographicLib C++ library; rather, this installs a *re-implementation*
in Python 3 of some of the algorithms present in the GeographicLib C++
library.