Preferences dialog, FlightGear settings: add tooltips to the labels
* ffgo/gui/configwindow.py (ConfigWindow.widgetFG): before this commit,
tooltips in the FlightGear settings tab were only created for the Entry
widgets. This commit adds tooltips to the corresponding Label widgets
too, so as to increase the chance the user realizes there is a tooltip.
I believe this actually restores the behavior in effect before I remade
this dialog using the 'grid' layout manager.