Bugfix: prevent unwanted selection changes in the airport and aircraft choosers
* Before this commit, the airport and aircraft choosers suffered from
the following problem: if the associated search field was not empty,
some operations such as running FlightGear or changing the sort order in
the airport or aircraft chooser, could cause the selected item to be
changed based on the contents of the search field. This is because these
operations internally rebuild the Treeview widget (no MVC design pattern
here...), and this was done using the exact same algorithm as when the
contents of the search field is changed.

* This commit fixes the problem using a new 'preserveSelection'
parameter passed to the relevant methods of the IncrementalChooser
class: updateList(), setTreeData() and _autoUpdateTreeSelection().

* Since IncrementalChooser.updateList() was used as a callback for a
Tkinter variable observer, it wasn't possible to add a parameter without
breaking this use. Therefore, we now use a new wrapper method as the
callback (updateList_callback()), which calls updateList() with the
desired value for 'preserveSelection'. The value is currently hardcoded,
as it can't be conveniently passed due to the way Tkinter calls such
variable observers. This is not a problem, as only one value is needed
in this code path.